+++ apps/web/eslint.config.js
@@ -1,0 +1,3 @@
+import config from "@jianartifact/eslint-config";
+
+export default config;